added -start and -end options for processing subsets

+++ src/phase3.py
@@ -40,11 +40,11 @@ DICTIONARY_VALIDATOR_JAR = os.path.join(
     ROOT_DIR, "source/radx-data-dictionary-validator-app-1.3.4.jar"
 )
 METADATA_SPEC = os.path.join(ROOT_DIR, "reference/RADxMetadataSpecification.json")
-GLOBAL_HARMONIZED_DICT = os.path.join(ROOT_DIR, "reference/RADx-global_tier1_dict_2024-10-28.csv")
+GLOBAL_HARMONIZED_DICT = os.path.join(ROOT_DIR, "reference/RADx-global_tier1_dict_2025-01-14.csv")
 ERROR_FILE_NAME = "phase3_errors.csv"
 
 
-def phase3_checker(include_dirs, exclude_dirs):
+def phase3_checker(include_dirs, exclude_dirs, start, end):
     """
     Executes the phase 3 checks on the specified directories.
 
@@ -92,8 +92,9 @@ def phase3_checker(include_dirs, exclude_dirs):
             os.remove(json_file)
 
         # Remove the origcopy and transformcopy directories from a previous run
-        shutil.rmtree(origcopy_dir, ignore_errors=True)
-        shutil.rmtree(transformcopy_dir, ignore_errors=True)
+        if start == 0:
+            shutil.rmtree(origcopy_dir, ignore_errors=True)
+            shutil.rmtree(transformcopy_dir, ignore_errors=True)
 
         print(f"processing: {directory}", end="")
 
@@ -109,45 +110,50 @@ def phase3_checker(include_dirs, exclude_dirs):
             continue
 
         # Check metadata file for correct format and information
-        for file in glob.glob(os.path.join(work_dir, "rad_*_*-*_*_META_origcopy.csv")):
-            error = utils.check_origcopy_meta_file(file, error_messages)
+        meta_files = sorted(glob.glob(os.path.join(work_dir, "rad_*_*-*_*_META_origcopy.csv")))
+        for meta_file in meta_files[start:end]:
+            error = utils.check_origcopy_meta_file(meta_file, error_messages)
             if error:
                 utils.save_error_file(error_messages, error_file)
 
         if utils.handle_errors_and_continue(error_file, error_messages):
             continue
 
-        # Run validation suite on the origcopy files
-        if not is_valid_data(work_dir, "origcopy", error_file, error_messages):
+
+        if not is_valid_data(work_dir, "origcopy", start, end, error_file, error_messages):
             continue
 
-        step0(work_dir, origcopy_dir)
-        step1(work_dir, transformcopy_dir)
-        step2(work_dir, transformcopy_dir, GLOBAL_HARMONIZED_DICT)
-        step3(transformcopy_dir)
-        step4(work_dir, transformcopy_dir)
+        # Run validation suite on the origcopy files
+        step0(work_dir, start, end, origcopy_dir)
+        step1(work_dir, start, end, transformcopy_dir)
+        step2(work_dir, start, end, transformcopy_dir, GLOBAL_HARMONIZED_DICT)
+        step3(transformcopy_dir, start, end)
+        step4(work_dir, start, end, transformcopy_dir)
 
         # Run validation suite on the transformcopy files
-        if not is_valid_data(transformcopy_dir, "transformcopy", error_file, error_messages):
+        # TODO do the start and end indices apply here? Not the same as for the origcopy files!
+        if not is_valid_data(transformcopy_dir, "transformcopy", start, end, error_file, error_messages):
             continue
 
         # Remove the metadata .csv files, since they has been replaced by .json files.
-        for meta_file in glob.glob(
-            os.path.join(transformcopy_dir, "rad_*_*-*_*_META_transformcopy.csv")
-        ):
+        # TODO do the start and end indices apply here? Not the same as for the origcopy files!
+        meta_files = sorted(glob.glob(os.path.join(transformcopy_dir, "rad_*_*-*_*_META_transformcopy.csv")))
+        for meta_file in meta_files[start:end]:
             os.remove(meta_file)
 
-        origcopies, transformcopies = utils.final_consistency_check(
-            preorigcopy_dir, origcopy_dir, transformcopy_dir, error_messages
-        )
+        num_files = len(glob.glob(os.path.join(work_dir, "rad_*_*-*_*_META_origcopy.csv")))
+        if end >= num_files:
+            origcopies, transformcopies = utils.final_consistency_check(
+               preorigcopy_dir, origcopy_dir, transformcopy_dir, error_messages
+            )
 
-        if utils.handle_errors_and_continue(error_file, error_messages):
-            continue
+            if utils.handle_errors_and_continue(error_file, error_messages):
+                continue
 
-        print(f" - {origcopies} origcopy and {transformcopies} transformcopy files created")
+            print(f" - {origcopies} origcopy and {transformcopies} transformcopy files created")
 
 
-def is_valid_data(work_dir, dir_type, error_file, error_messages):
+def is_valid_data(work_dir, dir_type, start, end, error_file, error_messages):
     """
     Validates the data, metadata, and dictionary for a given directory type.
 
@@ -167,22 +173,22 @@ def is_valid_data(work_dir, dir_type, error_file, error_messages):
     bool
         True if data validation succeeds, False otherwise.
     """
-    compile_metadata(work_dir, dir_type, error_messages)
+    compile_metadata(work_dir, dir_type, start, end, error_messages)
     if utils.handle_errors_and_continue(error_file, error_messages):
         return False
 
-    validate_dictionary(work_dir, dir_type, error_messages)
+    validate_dictionary(work_dir, dir_type, start, end, error_messages)
     if utils.handle_errors_and_continue(error_file, error_messages):
         return False
 
-    validate_metadata(work_dir, dir_type, error_messages)
+    validate_metadata(work_dir, dir_type, start, end, error_messages)
     if utils.handle_errors_and_continue(error_file, error_messages):
         return False
 
     return True
 
 
-def compile_metadata(work_dir, file_type, error_messages):
+def compile_metadata(work_dir, file_type, start, end, error_messages):
     """
     Compiles metadata CSV files into JSON format using an external compiler tool.
 
@@ -203,7 +209,13 @@ def compile_metadata(work_dir, file_type, error_messages):
     any_error = False
 
     # Compile metadata CSV files to JSON format
-    for meta_file in glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_META_{file_type}.csv")):
+#    for meta_file in glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_META_{file_type}.csv")):
+    meta_files = sorted(glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_META_{file_type}.csv")))
+    # TODO, if transformcopy, the start and end indices are not correct
+    if file_type == "origcopy":
+        meta_files = meta_files[start:end]
+        
+    for meta_file in meta_files:
         command = (
             f"java -jar {METADATA_COMPILER_JAR} -c {meta_file} -o {work_dir} -t {METADATA_SPEC}"
         )
@@ -222,7 +234,7 @@ def compile_metadata(work_dir, file_type, error_messages):
     return any_error
 
 
-def validate_dictionary(work_dir, file_type, error_messages):
+def validate_dictionary(work_dir, file_type, start, end, error_messages):
     """
     Validates dictionary CSV files using an external validator tool.
 
@@ -243,7 +255,11 @@ def validate_dictionary(work_dir, file_type, error_messages):
     any_error = False
 
     # Validate dictionary CSV file
-    for dict_file in glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_DICT_{file_type}.csv")):
+    dict_files = sorted(glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_DICT_{file_type}.csv")))
+    if file_type == "origcopy":
+        dict_files = dict_files[start:end]
+
+    for dict_file in dict_files:
         command = f"java -jar {DICTIONARY_VALIDATOR_JAR} --in={dict_file}"
         stderr = ""
         try:
@@ -260,7 +276,7 @@ def validate_dictionary(work_dir, file_type, error_messages):
     return any_error
 
 
-def validate_metadata(work_dir, file_type, error_messages):
+def validate_metadata(work_dir, file_type, start, end, error_messages):
     """
     Validates metadata files using an external validator tool.
 
@@ -280,8 +296,12 @@ def validate_metadata(work_dir, file_type, error_messages):
     """
     any_error = False
 
-    # Validate metadata files
-    for dict_file in glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_DICT_{file_type}.csv")):
+    # Validate metadata files:
+    dict_files = sorted(glob.glob(os.path.join(work_dir, f"rad_*_*-*_*_DICT_{file_type}.csv")))
+    if file_type == "origcopy":
+        dict_files = dict_files[start:end]
+
+    for dict_file in dict_files:
         data_file = dict_file.replace(f"DICT_{file_type}.csv", f"DATA_{file_type}.csv")
         meta_file = dict_file.replace(f"DICT_{file_type}.csv", f"META_{file_type}.json")
         command = (
@@ -303,7 +323,7 @@ def validate_metadata(work_dir, file_type, error_messages):
     return any_error
 
 
-def step0(work_dir, origcopy_dir):
+def step0(work_dir, start, end, origcopy_dir):
     """
     Prepares origcopy files by copying them from the working directory to the origcopy directory.
 
@@ -326,21 +346,20 @@ def step0(work_dir, origcopy_dir):
     # Copy origcopy files from the work directory to the origcopy directory
     os.makedirs(origcopy_dir, exist_ok=True)
 
-    for work_file in glob.glob(os.path.join(work_dir, "rad_*_*-*_*_*_origcopy.csv")):
-        origcopy_file_name = os.path.basename(work_file)
-        # Use JSON file instead of the CSV file for metadata
-        if work_file.endswith("META_origcopy.csv"):
-            work_file = work_file.replace("META_origcopy.csv", "META_origcopy.json")
-        if origcopy_file_name.endswith("META_origcopy.csv"):
-            origcopy_file_name = origcopy_file_name.replace(
-                "META_origcopy.csv", "META_origcopy.json"
-            )
+    dict_files = sorted(glob.glob(os.path.join(work_dir, "rad_*_*-*_*_DICT_origcopy.csv")))
+    for dict_file in dict_files[start:end]:
+        dict_file_name = os.path.basename(dict_file)
+        data_file_name = dict_file_name.replace("_DICT_origcopy.csv", "_DATA_origcopy.csv")
+        meta_file_name = dict_file_name.replace("_DICT_origcopy.csv", "_META_origcopy.json")
+        input_files_names = [dict_file_name, data_file_name, meta_file_name]
+        
+        for input_file_name in input_files_names:
+            work_file = os.path.join(work_dir, input_file_name)
+            origcopy_file = os.path.join(origcopy_dir, input_file_name)
+            shutil.copyfile(work_file, origcopy_file)
 
-        origcopy_file = os.path.join(origcopy_dir, origcopy_file_name)
-        shutil.copyfile(work_file, origcopy_file)
 
-
-def step1(work_dir, transformcopy_dir):
+def step1(work_dir, start, end, transformcopy_dir):
     """
     Converts origcopy data files into transformcopy format.
 
@@ -356,7 +375,9 @@ def step1(work_dir, transformcopy_dir):
     None
     """
     primary_key = "Id"
-    for dict_file in glob.glob(os.path.join(work_dir, "rad_*_*-*_*_DICT_origcopy.csv")):
+    dict_files = sorted(glob.glob(os.path.join(work_dir, "rad_*_*-*_*_DICT_origcopy.csv")))
+
+    for dict_file in dict_files[start:end]:
         if utils.contains_min_cdes(dict_file, primary_key):
             os.makedirs(transformcopy_dir, exist_ok=True)
             # Convert data file to transform copy format
@@ -370,7 +391,7 @@ def step1(work_dir, transformcopy_dir):
             data.to_csv(transformed_data_file, index=False)
 
 
-def step2(work_dir, transformcopy_dir, global_harmonized_dict):
+def step2(work_dir, start, end, transformcopy_dir, global_harmonized_dict):
     """
     Converts origcopy dictionary files into global transformcopy format.
 
@@ -388,7 +409,9 @@ def step2(work_dir, transformcopy_dir, global_harmonized_dict):
     None
     """
     primary_key = "Id"
-    for dict_file in glob.glob(os.path.join(work_dir, "rad_*_*-*_*_DICT_origcopy.csv")):
+    dict_files = sorted(glob.glob(os.path.join(work_dir, "rad_*_*-*_*_DICT_origcopy.csv")))
+    
+    for dict_file in dict_files[start:end]:
         if utils.contains_min_cdes(dict_file, primary_key):
             dict_file_name = os.path.basename(dict_file)
             transformed_dict_file = os.path.join(transformcopy_dir, dict_file_name)
@@ -399,7 +422,7 @@ def step2(work_dir, transformcopy_dir, global_harmonized_dict):
             dictionary.to_csv(transformed_dict_file, index=False)
 
 
-def step3(transformcopy_dir):
+def step3(transformcopy_dir, start, end):
     """
     Matches the global dictionary to the transformcopy data files.
 
@@ -412,15 +435,14 @@ def step3(transformcopy_dir):
     -------
     None
     """
-    for dict_file in glob.glob(
-        os.path.join(transformcopy_dir, "rad_*_*-*_*_DICT_transformcopy.csv")
-    ):
+    dict_files = sorted(glob.glob(os.path.join(transformcopy_dir, "rad_*_*-*_*_DICT_transformcopy.csv")))
+    for dict_file in dict_files[start:end]:
         data_file = dict_file.replace("_DICT_transformcopy.csv", "_DATA_transformcopy.csv")
         dictionary = utils.global_data_dict_matcher(data_file, dict_file)
         dictionary.to_csv(dict_file, index=False)
 
 
-def step4(work_dir, transformcopy_dir):
+def step4(work_dir, start, end, transformcopy_dir):
     """
     Copies and updates metadata for transformcopy files.
 
@@ -435,9 +457,12 @@ def step4(work_dir, transformcopy_dir):
     -------
     None
     """
-    for data_file in glob.glob(
-        os.path.join(transformcopy_dir, "rad_*_*-*_*_DATA_transformcopy.csv")
-    ):
+    # for data_file in glob.glob(
+    #     os.path.join(transformcopy_dir, "rad_*_*-*_*_DATA_transformcopy.csv")
+    # )
+    data_files = sorted(glob.glob(os.path.join(transformcopy_dir, "rad_*_*-*_*_DATA_transformcopy.csv")))
+    #for data_file in data_files[start:end]:
+    for data_file in data_files:
         # For the DATA files in the transformcopy directory, copy the META file
         data_file_name = os.path.basename(data_file)
         meta_file_name = data_file_name.replace("_DATA_transformcopy.csv", "_META_origcopy.csv")
@@ -455,7 +480,7 @@ def step4(work_dir, transformcopy_dir):
         utils.update_sha256_digest(data_file, meta_transformcopy_file)
 
 
-def main(include, exclude):
+def main(include, exclude, start, end):
     """
     Main function to execute the phase3_checker with command-line arguments.
 
@@ -490,8 +515,8 @@ def main(include, exclude):
 
     print()
 
-    # Run phase 2 check
-    phase3_checker(include, exclude)
+    # adjust start index to zero-based indexing
+    phase3_checker(include, exclude, start-1, end)
 
 
 if __name__ == "__main__":
@@ -513,9 +538,21 @@ if __name__ == "__main__":
         required=False,
         help="Comma-separated list of projects to exclude.",
     )
+    parser.add_argument(
+        "-start",
+        type=int,
+        default=1,
+        help="Index of the first file to process (default: 1)",
+    )
+    parser.add_argument(
+        "-end",
+        type=int,
+        default=sys.maxsize,
+        help="Index of the last file to process (default: no limit)",
+    )
 
     # Parse the arguments
     args = parser.parse_args()
 
     # Call the main function with the parsed arguments
-    main(args.include, args.exclude)
+    main(args.include, args.exclude, args.start, args.end)
